+++ RELEASE-NOTES.md
@@ -43,6 +43,7 @@
 ## Fixes and improvements
 * Fixed failing snow sql command when snowflake.yml is invalid and query has no templating.
 * Fix JSON serialization for `Decimal`, `time` and `binary`.
+* Added `--format=JSON_EXT` option to return JSON objects as proper JSON structures rather than strings.
 * Refactored Streamlit app deployment (using `FROM <stage>` syntax); removed deprecated Streamlit features